+++ server/impl/src/main/java/com/walmartlabs/concord/server/agent/AgentResource.java
@@ -20,7 +20,6 @@ package com.walmartlabs.concord.server.agent;
  * =====
  */
 
-import com.walmartlabs.concord.common.ConfigurationUtils;
 import com.walmartlabs.concord.server.AgentWorkerUtils;
 import com.walmartlabs.concord.server.sdk.metrics.WithTimer;
 import io.swagger.annotations.Api;
@@ -40,8 +39,6 @@ import javax.ws.rs.core.MediaType;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Map;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Named