+++ server/impl/src/main/java/com/walmartlabs/concord/server/events/AbstractEventResource.java
@@ -24,7 +24,6 @@ import com.google.common.util.concurrent.SettableFuture;
 import com.walmartlabs.concord.sdk.Constants;
 import com.walmartlabs.concord.server.cfg.ExternalEventsConfiguration;
 import com.walmartlabs.concord.server.cfg.TriggersConfiguration;
-import com.walmartlabs.concord.server.jooq.tables.UserRoles;
 import com.walmartlabs.concord.server.org.project.ProjectDao;
 import com.walmartlabs.concord.server.org.project.RepositoryDao;
 import com.walmartlabs.concord.server.org.triggers.TriggerEntry;